requirements.txt
Django==3.2
django-extensions==2.1.3
djangorestframework==3.12.4
isort==4.3.4
mccabe==0.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
six==1.11.0
